Move utility.py to directory
