+++ src/incognita/utility/utility.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from functools import wraps
-import json
 from pathlib import Path
 import time
 from typing import TYPE_CHECKING
