+++ tests/test_utility.py
@@ -8,7 +8,7 @@ import pytest
 from incognita import utility
 from incognita.data import ons_pd
 from incognita.logger import logger
-from incognita.utility import time_function
+from incognita.utility.utility import time_function
 
 
 class ONSPostcodeDirectoryStub(ons_pd.ONSPostcodeDirectory):
